+++ user/models.py
@@ -1,10 +1,11 @@
-import mailbox
 from django.db import models
 from django.contrib.auth.models import User
-from route.models import Company
+# from django.db.models.signals import post_save
+# from django.dispatch import receiver
+from company.models import Company
 
 # Create your models here.
-class UserProfile(models.Model):
+class UserInfo(models.Model):
     user=models.OneToOneField(
         User,
         on_delete=models.CASCADE,
@@ -17,3 +18,12 @@ class UserProfile(models.Model):
     company=models.ForeignKey(Company,on_delete=models.SET_NULL,blank=True,verbose_name="所在公司")
     wage=models.DecimalField(default=0.0,max_digits=None,decimal_places=2,blank=True,verbose_name="钱")
 
+# @receiver(post_save, sender=User)
+# def create_user_profile(sender, instance, created, **kwargs):
+#     if created:
+#         UserInfo.objects.create(user=instance)
+
+
+# @receiver(post_save, sender=User)
+# def save_user_profile(sender, instance, **kwargs):
+#     instance.profile.save()
